Cleanup of overlay text function and minor code cleanup
